+++ src/main/java/org/cloudsimplus/vmplacementgroup/VmPlacementGroupAffinityType.java
@@ -21,14 +21,14 @@
  *     You should have received a copy of the GNU General Public License
  *     along with CloudSim Plus. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.cloudbus.cloudsim.vms.vmplacementgroup;
+package org.cloudsimplus.vmplacementgroup;
 
-import org.cloudbus.cloudsim.allocationpolicies.vmplacementgroups.VmAllocationPolicyBestFitWithPlacementGroups_LRRL;
-import org.cloudbus.cloudsim.allocationpolicies.vmplacementgroups.VmAllocationPolicyChicSchedAllPack;
+import org.cloudsimplus.vmplacementgroups.VmAllocationPolicyBestFitWithPlacementGroups_LRRL;
+import org.cloudsimplus.vmplacementgroups.VmAllocationPolicyChicSchedAllPack;
 
 /**
- * An enumerator for the affinity type of a {@link VmPlacementGroup}. Depending on the 
- * {@link VmPlacementGroupScope} of the group, <i>Affinity</i> corresponds to the case where we 
+ * An enumerator for the affinity type of a {@link VmPlacementGroup}. Depending on the
+ * {@link VmPlacementGroupScope} of the group, <i>Affinity</i> corresponds to the case where we
  * want pack the VMs together (e.g., under a switch or inside a host), while <i>Anti-affinity</i>
  * corresponds to the opposite case where we want to spread out the VMs (e.g., under different
  * switches or in different hosts).
@@ -43,12 +43,12 @@ import org.cloudbus.cloudsim.allocationpolicies.vmplacementgroups.VmAllocationPo
  * @author Pavlos Maniotis
  */
 public enum VmPlacementGroupAffinityType {
-	
+
 	/**
 	 * Denotes the desire to pack the VMs together
 	 */
 	AFFINITY,
-	
+
 	/**
 	 * Denotes the desire to spread out the VMs
 	 */
